routing opstellen
voor het starten van het project #121 moesten we de routing aanpassen @ J3SS3HVA had eerder al wat aanpassingen gemaakt.
